+++ services/resources-getter.js
@@ -39,6 +39,10 @@ function ResourcesGetter(model, opts, params) {
       associationsForQuery);
   })();
 
+  var searchBuilder = new SearchBuilder(model, opts, params,
+      fieldNamesRequested);
+  var hasSmartFieldSearch = false;
+
   function handleFilterParams() {
     var where = {};
     var conditions = [];
@@ -67,8 +71,7 @@ function ResourcesGetter(model, opts, params) {
     where[OPERATORS.AND] = [];
 
     if (params.search) {
-      where[OPERATORS.AND].push(new SearchBuilder(model, opts, params,
-          fieldNamesRequested).perform());
+      where[OPERATORS.AND].push(searchBuilder.perform());
     }
 
     if (params.filter) {
@@ -104,12 +107,19 @@ function ResourcesGetter(model, opts, params) {
           try {
             field.search(countOpts, params.search);
             field.search(findAllOpts, params.search);
+            hasSmartFieldSearch = true;
           } catch (error) {
             Interface.logger.error('Cannot search properly on Smart Field ' +
               field.field, error);
           }
         }
       });
+
+      var fieldsSearched = searchBuilder.getFieldsSearched();
+      if (fieldsSearched.length === 0 && !hasSmartFieldSearch) {
+        // NOTICE: No search condition has been set in the queryis possible on the current model,
+        return [0, []];
+      }
     }
 
     if (segmentScope) {
@@ -154,6 +164,12 @@ function ResourcesGetter(model, opts, params) {
     return getSegmentCondition()
       .then(getAndCountRecords)
       .spread(function (count, records) {
+        var fieldsSearched = null;
+
+        if (params.search) {
+          fieldsSearched = searchBuilder.getFieldsSearched();
+        }
+
         if (schema.isCompositePrimary) {
           records.forEach(function (record) {
             record.forestCompositePrimary =
@@ -161,7 +177,8 @@ function ResourcesGetter(model, opts, params) {
                 .createCompositePrimary();
           });
         }
-        return [count, records];
+
+        return [records, count, fieldsSearched];
       });
   };
 }