+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Search - Display highlighted matches on table view when searching.
+
+### Fixed
+- Search - Fix the search on Enum fields.
+- Search - Return empty result instead of the whole list of records if no field can be searched on the collection.
 
 ## RELEASE 2.8.7 - 2018-05-18
 ### Fixed